Add daily color change
Rather than randomly generating start/end colors for the puzzle
on refresh, we pull from a DB table storing the colors for the
day. We have a cron job set those to random colors once daily.
This way, everyone gets the same colors and we follow the
wordle-like "daily puzzle" format.